% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coast.R
\name{coast}
\alias{coast}
\alias{coastline}
\title{Display Gulf Coast}
\usage{
coast(col = "grey80", border = "grey50", lwd = 0.4, ...)

coastline(col = "grey80", border = "grey50", lwd = 0.4, ...)
}
\arguments{
\item{col}{The colour to be used for in-filling the landmass. If}

\item{border}{The border colour to be used for displaying the coastline.}

\item{lwd}{Border line width (see \code{\link[graphics]{par}}).}

\item{\dots}{Further arguments to be passed to the \code{\link[graphics]{polygon}} function.}

\item{resolution}{Character string specifying the resolution of the
coastline, topographical and bathymetric data file to be loaded. It may take
the following values: \sQuote{low} (low resolution), \sQuote{intermediate} (intermediate resolution),
\sQuote{high} (high resolution), or \sQuote{full} (full resolution).}
}
\value{
A spatial object containing the coastline coordinates is returned invisibly.
}
\description{
Displays the coast and landmass surrounding the Gulf of Saint Lawrence.
}
\section{Functions}{
\itemize{
\item \code{coastline()}: Alias for \code{coast}

}}
\examples{
# Display simple map:
plot(c(-66.5, -60), c(45, 49), type = "n")
coast()

# Read coastline data:
x <- coast()

}
\seealso{
\code{\link[graphics]{polygon}}
}
