% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deg2grid.R
\name{deg2grid}
\alias{deg2grid}
\alias{grid2deg}
\title{Coordinates and Grids}
\usage{
deg2grid(
  x,
  y,
  quarter.grid = FALSE,
  xref = -66 - 1/3,
  yref = 45,
  dx = 1/6,
  dy = 1/6,
  reference = "HP23"
)

grid2deg(
  x,
  xref = -66 - 1/3,
  yref = 45,
  dx = 1/6,
  dy = 1/6,
  reference = "HP23",
  vertices = FALSE
)
}
\arguments{
\item{x}{Horizontal coordinate vector (e.g. decimal longitude) or grid name.}

\item{y}{Vertical coordinate vector (e.g. decimal latitude).}

\item{quarter.grid}{A logical value specifying whether the grid is to include northing and easting qualifiers.}

\item{xref}{A numerical value specifying the horizontal reference coordinate. Reference coordinates specify the location of the the lower-left corner of reference grid.}

\item{yref}{A numerical value specifying the vertical reference coordinate. Reference coordinates specify the location of the the lower-left corner of reference grid.}

\item{dx}{A numerical value specifying the horizontal width of each regular grid.}

\item{dy}{A numerical value specifying the vertical height of each regular grid.}

\item{reference}{A four-character string of the form XXYY, where XX are two letters and YY are two digits.}
}
\description{
Functions to assign a grid label(s) given coordinates, and vice versa.
}
\section{Functions}{
\itemize{
\item \code{deg2grid}: Returns a grid name for a given coordinate(s).

\item \code{grid2deg}: Returns the corner coordinates of a grid.
}}

\examples{
deg2grid(-64, 47) # Returns "HD37".
grid2deg("HD37")  # Return corner coordinates for "HD37".
grid2deg("HD37", vertices = TRUE)  # Return vertex coordinates for "HD37".

# Generate 100 random coordinates:
lon <- -66 + 6 * runif(100)
lat <- 45 + 4 * runif(100)
deg2grid(lon, lat)

}
\seealso{
\code{\link{plot.grid}}
}
