% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deg2str.R
\name{deg2str}
\alias{deg2str}
\title{Coordinate Conversion Functions}
\usage{
deg2str(
  deg = NULL,
  latitude = NULL,
  longitude = NULL,
  clip = TRUE,
  northing = TRUE,
  easting = TRUE,
  as.list = FALSE
)
}
\arguments{
\item{deg}{A numeric vector of coordinates in decimal degree format.}

\item{longitude, latitude}{Numeric decimal degree coordinate value(s). If \code{northing = TRUE} or
\code{easting = TRUE} then the corresponding Northing (i.e. "N" or "S") or Easting character
(i.e. "E" or "W") is included in the output character string.}

\item{clip}{A logical value. Whether to strip away coordinate seconds or
minutes and seconds when they have zero values. The resulting output strings
are of an abridged form. The default is \code{TRUE}.}

\item{northing, easting}{Logical value specifying whether to include a Northing (i.e. "N" or "S")
or Easting (i.e. "E" or "W") character in the output string when \code{longitude}
or \code{latitude} are specified.}

\item{as.list}{Logical value specifying whether to return the output as a list object. The default is \code{FALSE}.}
}
\value{
If a single argument is provided and as.list is FALSE, then the
returned output is a character string vector. Otherwise, the output is
returned as a list with fields \code{longitude} and \code{latitude}.
}
\description{
Convert a decimal-degree coordinate to a formatted character string suitable for printing.
}
\examples{
# Convert generic coordinate:
deg2str(c(45, 47.5))
deg2str(c(45, 47.5), clip = FALSE)

# Convert latitude coordinate, which adds the northing character to the output:
deg2str(lat = c(45, 47.5))
deg2str(long = c(-66, -66.5))

# Convert vectors of lat-long coordinates, which are returned as a list:
deg2str(latitude = 45:47, longitude = -64:-66, clip = FALSE)

}
\seealso{
loran2deg deg2grid deg2grid
}
