% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depth.R
\name{depth}
\alias{depth}
\alias{depth.default}
\alias{depth.list}
\title{Estimate Water Depth}
\usage{
depth(x, ...)

\method{depth}{default}(longitude, latitude, units = "m")

\method{depth}{list}(x, ...)
}
\arguments{
\item{longitude, latitude}{Numerical vector of longitudes in decimal degree format.}

\item{units}{Depth units to be returned. May be in meters (\code{units = "meter"}), feet (\code{units
= "ft"} or \code{units = "feet"}) or fathoms (\code{units = "fth"} or (\code{units = "fathoms"}).}
}
\value{
Returns a numerical vector the same size as \code{latitude} and \code{longitude} containing
the water depth at the specified coordinates. Coordinates on land have negative values.
}
\description{
Returns an linearly interpolated estimate of water depth for a given a set
of coordinates, based on a reference bathymetry map.
}
\details{
The sign of \code{longitude} has no influence on the result.
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Default \code{depth} function.

\item \code{list}: Depth function for a list obejct.
}}

\examples{
   # Find the water depth at a single point:
   depth(63.8, 47.05)

   # Find water depth for multiple points:
   lat <- c(48, 47, 46.5)
   long <- c(-64, -61.5, -62)
   depth(long, lat)

}
