% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishing.zone.R
\name{fishing.zone}
\alias{fishing.zone}
\title{Determine Fishing Zone}
\usage{
fishing.zone(longitude, latitude, species, ...)
}
\arguments{
\item{longitude}{Numerical vector of longitudes in decimal degree format.}

\item{latitude}{Numerical vector of latitudes in decimal degree format.}

\item{species}{Numerical scalar containing a species code.}

\item{...}{Further arguments passed on to read.gulf.spatial}
}
\description{
Returns a fishing zone given a set of of coordinates.
}
\examples{
# Find the fishing.zone at a single point:
fishing.zone(-63.8, 47.05, species = 2526)

# Find strata values for multiple points:
lat <- c(48, 47, 46.5)
long <- c(-64, -61.5, -62)
fishing.zone(long, lat, species = 2526)

# Read snow crab set card for 2020 and determine fishing zone:
x <- read.scsset(2020)
fishing.zone(lon(x), lat(x), species = 2526)

# Lobster fishing areas of snow crab survey tows.
fishing.zone(lon(x), lat(x), species = 2550)
}
