% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fishing.zone.vertices}
\alias{fishing.zone.vertices}
\title{Vertices of fishing zones.}
\format{
A simple feature data frame with 116 rows and 5 variables:
\describe{
  \item{type}{type of simple feature}
  \item{species.code}{species code of fishing zone}
  \item{region}{region of fishing zone}
  \item{label}{label of fishing zone}
  \item{geometry}{simple feature geometry}
}
}
\source{
\url{https://laws-lois.justice.gc.ca/eng/regulations/SOR-86-21/index.html}
}
\usage{
fishing.zone.vertices
}
\description{
A dataset containing the vertices of fishing zones.
}
\keyword{datasets}
