% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gulf.spatial-package.R
\docType{package}
\name{gulf.spatial-package}
\alias{gulf.spatial}
\alias{gulf.spatial-package}
\title{gulf.spatial: DFO Gulf Region Spatial Functions}
\description{
Collection of spatial functions for use in DFO Gulf Region, the southern Gulf of Saint Lawrence.
}
\author{
\strong{Maintainer}: Tobie Surette \email{Tobie.Surette@dfo-mpo.gc.ca}

Authors:
\itemize{
  \item Tobie Surette \email{Tobie.Surette@dfo-mpo.gc.ca}
  \item Daniel Ricard \email{Daniel.Ricard@dfo-mpo.gc.ca}
}

}
\keyword{internal}
