% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gulf.spatial.R
\docType{package}
\name{gulf.spatial}
\alias{gulf.spatial}
\title{gulf.spatial: sGSL Utility Functions}
\description{
A package of spatial functions for use in the southern Gulf of Saint Lawrence
fisheries' science.
}
\section{\code{gulf.spatial} functions}{


\code{\link{match}} matches the rows between data frames.
}

