% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loran2deg.R
\name{loran2deg}
\alias{loran2deg}
\alias{loran}
\alias{as.loran}
\alias{is.loran}
\title{Loran Coordinate Conversion}
\usage{
loran2deg(x, y)

as.loran(x)

is.loran(x)
}
\arguments{
\item{x}{Numeric vector of Loran X coordinates.}

\item{y}{Numeric vector of Loran Y coordinates.}
}
\value{
Data frame with fields \code{long} and \code{lat} in decimal degree format with the same
dimensions as \code{x} and \code{y}.
}
\description{
Function to convert Loran coordinates format to decimal latitude-longitude degree format.
}
\section{Functions}{
\itemize{
\item \code{as.loran}: Clean up loran coordinate strings.

\item \code{is.loran}: Determine if character string is a Loran-type coordinate.
}}

\examples{
# Check whether character string is a candidate loran coordinate:
is.loran(c("", "25.000", "0025500", "14456.123", "014512345", "14'456", "0000"))

# Loran coordinate clean-up function:
as.loran(c("", "25.000", "0025500", "14456.123", "014512345", "14'456", "0000"))

# Convert a Loran XY coordinate to decimal degree format:
loran2deg(14500, 29500)

# Convert a vector of Loran XY coordinates to decimal degree format:
loran2deg(seq(14000, 14500, len = 21), seq(29000, 29500, len = 21))

}
