% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.axis.R
\name{map.axis}
\alias{map.axis}
\title{Draw Axes on a Map}
\usage{
\method{map}{axis}(
  side = 1,
  at,
  major.interval = 1,
  major.cex = 0.6,
  minor.cex = 0.4,
  cex = 1,
  language = "english",
  ...
)
}
\arguments{
\item{side}{An integer(s) specifying which side(s) of the plot the axis is to be drawn on. The axis is placed
as follows: 1=below, 2=left, 3=above and 4=right.}

\item{at}{A vector in decimal degrees specifying the points at which tick-marks are to be drawn.}

\item{major.interval}{A numerical scalar specifying the modulus which define where major tick marks lie.
The default value is 1.}

\item{major.cex}{A numerical value specifying the major tick mark character expansion value.}

\item{minor.cex}{A numerical value specifying the minor tick mark character expansion value.}

\item{\dots}{Further parameters passed on to the \code{\link[graphics]{axis}} function.}
}
\description{
Draws axes on a map.
}
\seealso{
\code{\link{map}}, \code{\link[graphics]{axis}}
map.new()
map.axis()
}
