% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{map}
\alias{map}
\alias{map.default}
\alias{map.esonar}
\title{Map Gulf of St. Lawrence Features}
\usage{
map(x, ...)

\method{map}{default}(layer, add = TRUE, ...)

\method{map}{esonar}(x, variable, ...)
}
\arguments{
\item{layer}{Layer specification}
}
\description{
Draws southern Gulf of St. Lawrence geographical features.
}
\section{Methods (by class)}{
\itemize{
\item \code{map(default)}: Plot Gulf of Saint Lawrence base map.

\item \code{map(esonar)}: Plot \strong{eSonar} coordinate track data.

}}
\examples{
map()
map(layer = "strata", survey = "rv", region = "gulf")

}
\seealso{
\code{\link{read.gulf.spatial}}
}
