% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.grid.R
\name{plot.grid}
\alias{plot.grid}
\title{Coordinates and Grids}
\usage{
\method{plot}{grid}(x, label, col = "white", border = "black", ...)
}
\arguments{
\item{label}{Logical value or character strings specifying whether or what to display as labels when plotting grids.}

\item{col}{Background colour when plotting a grid using \code{plot.grid}.}

\item{border}{Border line colour when plotting a grid using \code{plot.grid}.}

\item{vertices}{Logical value specifiying whether to output the corner coordinates of each grid.}
}
\description{
Functions to assign a grid label(s) given coordinates, and vice versa.
}
\examples{
# Generate set of snow crab 10'x10' grids:
x <- expand.grid(seq(-66.5 + 1/12, -60 - 1/12, by  = 1/6),
                 seq(45.5-1/12, 49 + 1/12, by  = 1/6))
x$grid <- deg2grid(x[,1], x[,2])
x <- grid2deg(x$grid, vertices = TRUE) # Grid corner coordinates.
plot(c(-66.5, -60), c(45, 49), type = "n")
coast()
plot.grid(x$grid)

}
\seealso{
\code{\link{deg2grid}}
}
