% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.gulf.spatial.R
\name{read.gulf.spatial}
\alias{read.gulf.spatial}
\title{Read Spatial Data}
\usage{
read.gulf.spatial(layer, survey, project, ...)
}
\arguments{
\item{layer}{Character string specifying the layer to be read.}

\item{survey}{Character string specifying the name of the science survey to be read.}

\item{project}{Character string specifying the name of the science project to be read.}
}
\description{
Read southern Gulf spatial data files.
}
\section{Geographic Layers}{

\describe{
  \item{\code{altitude}, \code{bathymetry}, \code{dem}, \code{depth}}{Load altitude and bathymetry data.}
  \item{\code{borders}, \code{bounds}}{Load the bounds or borders of a specified area.}
  \item{\code{cities}, \code{towns}}{Load towns and cities.}
  \item{\code{fishing.zones}}{Load fishing zone polygons and borders.}
  \item{\code{kriging}}{Load polygons used in area abundance and biomass estimation.}
  \item{\code{stations}}{Load survey or project sampling stations.}
  \item{\code{strata}, \code{stratum}}{Load survey strata.}
  \item{\code{seaports}, \code{ports}}{Load sea ports.}
  \item{\code{geography}, \code{features}}{Load land and marine geographical features.}
}
}

\examples{
read.spatial(layer = "bounds", survey = "rv") # Read Septmeber survey boundaries.
read.spatial(layer = "bounds", survey = "sc") # Read snow crab survey boundaries.
read.spatial(layer = "stations", survey = "rv") # Read Septmeber survey stations.
read.spatial(layer = "stations", survey = "ns") # Read Northumberland Strait survey stations.

}
