% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.mif.R
\name{read.mif}
\alias{read.mif}
\title{Read MapInfo MIF Files}
\usage{
read.mif(file, print = FALSE)
}
\arguments{
\item{file}{File name.}

\item{print}{Logical value specifying whether to print information about graphic objects being read.}
}
\description{
Reads MapInfo Interchange Format (MIF) files.
}
