% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scs.survey.grid.R
\name{scs.survey.grid}
\alias{scs.survey.grid}
\title{Identify Survey Sampling Grid}
\usage{
scs.survey.grid(x, ...)
}
\arguments{
\item{x}{Data object.}

\item{longitude, latitude}{Vector of decimal}
}
\description{
Returns the survey sampling grid to which a coordinate or data observation belongs.
}
\examples{
# Single point:
scs.survey.grid(-63, 47.5)

# Vector of points:
s <- read.scsset(2020)
scs.survey.grid(lon(s), lat(s))
scs.survey.grid(s) # Use data directly.

# Return survey grid definitions:
scs.survey.grid()

}
