% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/station.R
\name{station}
\alias{station}
\alias{station.default}
\alias{station.nssset}
\title{Return a Station Number}
\usage{
station(x, ...)

\method{station}{default}(longitude, latitude, time, distance.tolerance = 1, time.tolerance, ...)

\method{station}{nssset}(x, tolerance = NULL, method)
}
\arguments{
\item{longitude, latitude}{Coordinates in decimal degrees.}

\item{time}{Numeric or (\sQuote{POSIX}) time objects.}

\item{distance.tolerance}{Numeric value specifying the distance (in kilometers) below which a pair of points
is considered to belong to the same group or station.}

\item{time.tolerance}{Numeric value specifying the time-difference below which a pair of points
is considered to belong to the same group or station. The units are in minutes
if the \code{time} argument belongs to a \sQuote{POSIX} class.}
}
\description{
Assign or retrieve a sampling station identifier.
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Assign a spatial location identifier based on spatio-temporal proximity.

\item \code{nssset}: Return Northumberland Strait survey sampling station identifier.
}}

