% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratum.R
\name{stratum}
\alias{stratum}
\title{Survey Spatial Strata}
\usage{
stratum(x, ...)
}
\arguments{
\item{x}{Object.}

\item{...}{Further arguments (not used).}

\item{longitude, latitude}{Numerical longitude and latitude value(s) in decimal degree format.}

\item{region}{Character string specifying the geographic region.}

\item{survey}{Character string specifying the research survey. This argument
is passed onto the \code{\link{stratum.info}} function.}
}
\value{
Returns a numerical vector the same size as \code{latitude} and \code{longitude}
        containing the survey strata numbers.
}
\description{
Determine the spatial survey stata corresponding to a set of coordinates.
}
\examples{
# Get stratum information:
stratum(region = "gulf", survey = "rv")

# Find the stratum at a single point:
stratum(-63.8, 47.05) # returns 429

# Find strata values for multiple points:
lat <- c(48, 47, 46.5)
long <- c(-64, -61.5, -62)
stratum(long, lat) # returns: c(422, 434, 401)

# Read a set card:
x <- read.scsset(year = 2010)
stratum(longitude(x), latitude(x))
}
