% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/substrate.R
\name{substrate}
\alias{substrate}
\title{Local Substrate Information}
\usage{
substrate(longitude, latitude)
}
\arguments{
\item{longitude, latitude}{Numerical vector of longitudes in decimal degree format.}
}
\value{
Returns character vector of substrate composition (e.g. sand, gravel, etc.)
the same length as the coordinate vectors.
}
\description{
Returns local sea-floor substrate information.
}
\examples{
   # Find the water depth at a single point:
   substrate(63.8, 47.05)

   # Find water depth for multiple points:
   lat <- c(48, 47, 46.5)
   long <- c(-64, -61.5, -62)
   substrate(long, lat)

}
\references{
Loring and Nota, 1973, https://gcgeo.gc.ca/geonetwork/metadata/eng/8c269a91-d3a2-4f49-943d-6b2401c42cba, accessed April 10, 2025
}
