% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey.design.R
\name{scs.stations}
\alias{scs.stations}
\alias{nss.stations}
\title{Generate Survey Sampling Stations}
\usage{
scs.stations(
  n,
  longitude,
  latitude,
  hole = NULL,
  alternates = 0,
  grids = FALSE
)

nss.stations(n, alternates = 0, stations, type)
}
\arguments{
\item{n}{Integer specifying the desired number of sampling stations.}

\item{longitude, latitude}{Numeric vector specifying, in decimal degrees, the
longitude coordinate of each vertex of the survey polygon, which specifies
the survey area. Multiple, disjoint polygons may be separated by NA values.
If not specified, then the snow crab survey polygon is used.}

\item{hole}{Logical vector. When multiple polygons are specified in
\code{longitude} and \code{latitude}, then \code{hole} may be used to
specify if some are \sQuote{holes}. The length of \code{hole} may be the
length of the coordinate vectors or the number of polygons defined within
them.}

\item{alternates}{Integer specifying the number of additional, alternate sampling stations
per grid to be generated.  The default value is zero.}

\item{grids}{Logical value specifying whether the grid coordinates are to be
included in the output for the snow crab survey. Alternatively, the output from a
previous call of \code{select.sc.stations} may be used. Specifically, the grid
definitions specified in the \code{grids} element of the output is used to generate
points within each element. Alternately, if the complete of grids which overlay
the sGSL survey area is dersired, set \code{grids = \sQuote{complete}}
or\code{\sQuote{all}}. The default value is \code{FALSE}.}

\item{stations}{Data frame specifying the set of survey stations to be sampled.}

\item{type}{Character string specifying the 'type' label to be assigned in the resulting sample.}
}
\value{
If \code{grids} is \code{FALSE}, a data frame is returned with
   sampling station coordinates (\code{longitude} and \code{latitude}), the
   station type (whether the station is a primary or alternate station) and a
   grid label which identifies from which grid the sample was drawn.

   If \code{grids} is \code{TRUE}, then a two-element list is returned. The
   first element contains the station table described above, while the second
   element contains a list defining the coordinates of each grid vertex as well
   as the grid label.
}
\description{
Returns a set of randomly-selected snow crab survey sampling stations.
             Each station is drawn from a square grid, a lattice which overlays the
             survey area. Longitude and latitude coordinates of the survey polygon's
             vertices are converted to UTM coordinates (NAD83). A square grid lattice is
             overlaid over the survey area. The dimensions of the grid vary according to
             number of required sampling stations and the area of the survey polygon.
}
\section{Functions}{
\itemize{
\item \code{scs.stations}: Snow crab survey station design.

\item \code{nss.stations}: Northumberland Strait survey station design.
}}

\examples{
# Snow crab survey:
x <- scs.stations(355) # Generate 355 survey stations.
x <- scs.stations(355, alternates = 2) # Include 2 alternate station in the output.
x <- scs.stations(355, alternates = 2, grids = TRUE) # Include grids in the export.

# Northumberland Strait survey:
x <- nss.stations(100) # Generate 100 survey stations.
x <- nss.stations(80, alternates = 2) # Generate 80 x survey stations plus 2 alternates sets
}
