% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.geotiff.R
\name{write.geotiff}
\alias{write.geotiff}
\title{Write Image Data as GeoTIFF}
\usage{
write.geotiff(z, file.name, longitude, latitude)
}
\arguments{
\item{z}{Two-dimensional matrix or array.}

\item{file.name}{File name.}

\item{longitude, latitude}{Vectors of regularly spaced coordinate that define the spatial extent of the image.
If \code{longitude}.}
}
\description{
Write Grid or Image to a GeoTIFF. GeoTIFF is a public domain metadata standard which allows georeferencing
             information to be embedded within a TIFF file. The potential additional information includes map projection,
             coordinate systems, ellipsoids, datums, and everything else necessary to establish the exact spatial reference
             for the file.
}
