% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinates.R
\name{longitude}
\alias{longitude}
\alias{latitude}
\alias{longitude.default}
\alias{latitude.default}
\alias{longitude.data.frame}
\alias{latitude.data.frame}
\alias{lon}
\alias{lat}
\alias{long}
\title{Retrieve Spatial Coordinates}
\usage{
longitude(x, ...)

latitude(x, ...)

\method{longitude}{default}(x, ...)

\method{latitude}{default}(x, ...)

\method{longitude}{data.frame}(x, ...)

\method{latitude}{data.frame}(x, ...)

lon(x, ...)

lat(x, ...)

long(x, ...)
}
\arguments{
\item{x}{Object containing coordinates. Coordinates can be extracted from list and
data frame objects if they either have defined \code{longitude} and \code{latitude}
methods already defined, or looking for \sQuote{longitude} or \sQuote{latitude}
in the field variable names. In the latter case, partial matching is used.}
}
\description{
Functions to retrieve coordinates from an object.
}
\section{Functions}{

\describe{
  \item{\code{longitude}, \code{latitude}}{Extract longitude or latitude coordinates from a data frame or list.}
  \item{\code{lon}, \code{long}}{Alias functions for \code{longitude}.}
  \item{\code{lat}}{Alias function for \code{latitude}.}
}
}

