% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deg2dmm.R
\name{deg2dmm}
\alias{deg2dmm}
\alias{dmm2deg}
\alias{deg2dms}
\alias{dms2deg}
\title{Latitude-Longitude Conversion}
\usage{
deg2dmm(x)

dmm2deg(x)

deg2dms(x)

dms2deg(x)
}
\arguments{
\item{x}{Numeric vector.}
}
\description{
Converts from one latitude-longitude numeric format to another.
}
\section{Functions}{
\itemize{
\item \code{dmm2deg}: Convert from degree-decimal minute to decimal degree format.

\item \code{deg2dms}: Convert from decimal degree to degree-minute-second format.

\item \code{dms2deg}: Convert from degree-minute-second to decimal degree format.
}}

\examples{
# Degree decimal-minute conversions:
deg2dmm(66.5) # Returns 6630.00
dmm2deg(6630) # Returns 66.5
deg2dmm(60+5*runif(100)) # Convert 100 random coordinates to DMM format.

# Degree-minute-second conversions:
deg2dms(66.5) # Returns 663000
dms2deg(663000) # Returns 66.5
deg2dms(60+5*runif(100)) # Convert 100 random coordinates to DMS format.

}
\seealso{
loran2deg deg2grid deg2grid

Convert from decimal degree to degree decimal-minute format.
}
