% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deg2grid.R
\name{deg2grid}
\alias{deg2grid}
\alias{grid2deg}
\alias{plot.grid}
\title{Coordinates and Grids}
\usage{
deg2grid(
  x,
  y,
  quarter.grid = FALSE,
  xref = -66 - 1/3,
  yref = 45,
  dx = 1/6,
  dy = 1/6,
  reference = "HP23"
)

grid2deg(
  x,
  xref = -66 - 1/3,
  yref = 45,
  dx = 1/6,
  dy = 1/6,
  reference = "HP23",
  vertices = FALSE
)

plot.grid(x, label = FALSE, col = "white", border = "black", ...)
}
\arguments{
\item{x}{Horizontal coordinate vector (e.g. decimal longitude) or grid name.}

\item{y}{Vertical coordinate vector (e.g. decimal latitude).}

\item{quarter.grid}{A logical value specifying whether the grid is to include northing and easting qualifiers.}

\item{xref}{A numerical value specifying the horizontal reference coordinate. Reference coordinates specify the location of the the lower-left corner of reference grid.}

\item{yref}{A numerical value specifying the vertical reference coordinate. Reference coordinates specify the location of the the lower-left corner of reference grid.}

\item{dx}{A numerical value specifying the horizontal width of each regular grid.}

\item{dy}{A numerical value specifying the vertical height of each regular grid.}

\item{reference}{A four-character string of the form XXYY, where XX are two letters and YY are two digits.}

\item{vertices}{Logical value specifiying whether to output the corner coordinates of each grid.}

\item{label}{Logical value specifying whether to display label inside the grids when using \code{plot.grid}.}

\item{col}{Background colour when plotting a grid using \code{plot.grid}.}

\item{border}{Border line colour when plotting a grid using \code{plot.grid}.}
}
\description{
Functions to assign a grid label(s) given coordinates, and vice versa.
}
\section{Functions}{

\describe{
  \item{\code{deg2grid}}{Returns a grid name for a given coordinate(s).}
  \item{\code{grid2deg}}{Returns the corner coordinates of a grid.}
  \item{\code{plot.grid}}{Graphically display grid.}
}
}

\examples{
deg2grid(-64, 47) # Returns "HD37".
grid2deg("HD37")  # Return corner coordinates for "HD37".
grid2deg("HD37", vertices = TRUE)  # Return vertex coordinates for "HD37".

# Generate 100 random coordinates:
lon <- -66 + 6 * runif(100)
lat <- 45 + 4 * runif(100)
deg2grid(lon, lat)

# Generate set of snow crab 10'x10' grids:
x <- expand.grid(seq(-66.5 + 1/12, -60 - 1/12, by  = 1/6),
                 seq(45.5-1/12, 49 + 1/12, by  = 1/6))
x$grid <- deg2grid(x[,1], x[,2])
x <- grid2deg(x$grid, vertices = TRUE) # Grid corner coordinates.
plot(c(-66.5, -60), c(45, 49), type = "n")
coast()
plot.grid(x$grid)

}
