% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deg2km.R
\name{deg2km}
\alias{deg2km}
\alias{km2deg}
\title{Degree and Kilometer Conversion}
\usage{
deg2km(x, y, long.ref, lat.ref, method = "utm")

km2deg(x, y, long.ref, lat.ref, method = "utm")
}
\arguments{
\item{x, y}{A numeric vector of longitude and latitude coordinates. Coordinates are assumed
to be in decimal degree format.}

\item{long.ref, lat.ref}{Reference coordinate from which the coordinate
distances will be calculated. Assumed to be in decimal degree format.}

\item{method}{Character string specifying the method to be used for calculating the distances
between coordinate points. The options are \sQuote{utm} which first converts the
coordinates into UTM NAD83 coordinates, and \sQuote{ellipsoid} which models the
Earth as an ellipsoid with major and minor radii of 6378.16 and 6356.775 kilomters,
respectively. The default is \sQuote{utm}.}
}
\value{
A data frame containing the converted coordinates.
}
\description{
Convert geographical coordinates to kilometers and vice versa.
}
\section{Functions}{
\itemize{
\item \code{km2deg}: Convert coordinates from kilometers to decimal lat-lon.
}}

\examples{
# Convert a decimal degree coordinate to kilometers format:
deg2km(-63.123, 47.56)

# Convert a vector of decimal degree coordinates to kilometer format:
deg2km(seq(-65, -63, len = 21), seq(46, 48, len = 21))

#Convert a coordinate in kilometer format to decimal degree format:
km2deg(-227.3639, 111.1331)

# Return the value of the reference coordinate:
km2deg(0, 0)

# Convert a vector of kilometer-format coordinates to decimal degree format:
km2deg(seq(0, -200, len = 21), seq(0, 100, len = 21))

}
\seealso{
loran2deg deg2dms deg2dmm deg2str
}
