% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishing.zone.R
\name{fishing.zone}
\alias{fishing.zone}
\alias{fishing.zone.default}
\title{Determine Fishing Zone}
\usage{
fishing.zone(x, ...)

\method{fishing.zone}{default}(longitude, latitude, species = NULL, region = "gulf", ...)
}
\arguments{
\item{longitude}{Numerical vector of longitudes in decimal degree format.}

\item{latitude}{Numerical vector of latitudes in decimal degree format.}

\item{species}{Numerical scalar containing a species code.}

\item{region}{Character value specifying the geographic region. This argument is passed onto the \code{\link[gulf]{fishing.zone.info}} function.}
}
\description{
Returns a fishing zone given a set of of coordinates.
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Returns a fishing zone for a set of coordinate points.
}}

\examples{
   # Find the fishing.zone at a single point:
   fishing.zone(-63.8, 47.05, species = 2526)

   # Find strata values for multiple points:
   lat <- c(48, 47, 46.5)
   long <- c(-64, -61.5, -62)
   fishing.zone(long, lat, species = 2526)

   # Read snow crab set card for 2013 and determine fishing zone:
   x <- read.gulf(year = 2013, survey = "sc")
   zone <- fishing.zone(longitude(x), latitude(x), species = 2526)

   # Read rv set card for 2013 and determine fishing zone:
   x <- read.gulf(year = 2013)
   zone <- fishing.zone(longitude(x), latitude(x), species = 60)
}
