% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishing.zone.R
\name{fishing.zone.default}
\alias{fishing.zone.default}
\title{Determine Fishing Zone}
\usage{
\method{fishing.zone}{default}(longitude, latitude, species, ...)
}
\arguments{
\item{longitude}{Numerical vector of longitudes in decimal degree format.}

\item{latitude}{Numerical vector of latitudes in decimal degree format.}

\item{species}{Numerical scalar containing a species code.}

\item{region}{Character value specifying the geographic region. This argument is passed onto the \code{\link[gulf]{fishing.zone.info}} function.}
}
\description{
Returns a fishing zone given a set of of coordinates.
}
\section{Methods (by class)}{
\itemize{
\item \code{fishing.zone(default)}: Returns a fishing zone for a set of coordinate points.

}}
\examples{
# Find the fishing.zone at a single point:
fishing.zone(-63.8, 47.05, species = 2526)

# Find strata values for multiple points:
lat <- c(48, 47, 46.5)
long <- c(-64, -61.5, -62)
fishing.zone(long, lat, species = 2526)

x <- read.scsset(2020)
fishing.zone(lon(x), lat(x), species = 2526) # Read snow crab set card for 2013 and determine fishing zone:
fishing.zone(lon(x), lat(x), species = 2550) # Lobster fishing areas of snow crab survey tows.
}
