% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locate.R
\name{locate}
\alias{locate}
\alias{locate.gulf.spatial}
\title{Locate Spatial Data Files}
\usage{
\method{locate}{gulf.spatial}(
  layer,
  survey,
  region,
  project,
  file.extensions = c("csv", "shp", "txt", "tab", "rda", "rdata", "mif"),
  resolution = "high",
  ...
)
}
\arguments{
\item{layer}{Character string specifying the layer to be read.}

\item{survey}{Character string specifying the name of the science survey to be read.}

\item{project}{Character string specifying the name of the science project to be read.}
}
\description{
Read southern Gulf spatial data files.
}
\section{Geographic Layers}{

\describe{
  \item{\code{altitude}, \code{bathymetry}, \code{dem}, \code{depth}}{Load altitude and bathymetry data.}
  \item{\code{borders}, \code{bounds}}{Load the bounds or borders of a specified area.}
  \item{\code{grids}}{Science survey spatial grids designs.}
  \item{\code{cities}}{Load city coordinates.}
  \item{\code{fishing.zones}}{Load fishing zone polygons and borders.}
  \item{\code{kriging}}{Load polygons used in area abundance and biomass estimation.}
  \item{\code{stations}}{Load survey or project sampling stations.}
  \item{\code{strata}, \code{stratum}}{Load survey strata.}
  \item{\code{seaports}, \code{ports}}{Load sea ports.}
  \item{\code{geography}, \code{features}}{Load land and marine geographical features.}
}
}

\examples{
locate.gulf.spatial(layer = "bounds", survey = "rv")   # Locate September survey boundaries.
locate.gulf.spatial(layer = "bounds", survey = "sc")   # Locate snow crab survey boundaries.
locate.gulf.spatial(layer = "stations", survey = "rv") # Locate September survey stations.
locate.gulf.spatial(layer = "stations", survey = "ns") # Locate Northumberland Strait survey stations.
locate.gulf.spatial("grids", survey = "ns")            # Locate Northumberland Strait survey grids.

}
