% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{map}
\alias{map}
\alias{map.default}
\title{Map Gulf of St. Lawrence Features}
\usage{
map(x, ...)

\method{map}{default}(x, layer, xlim, ylim, region = "gulf", add = TRUE, ...)
}
\arguments{
\item{layer}{Layer specification}
}
\description{
Draws southern Gulf of St. Lawrence geographical features.
}
\examples{
map()
map(layer = "strata", survey = "rv", region = "gulf")

}
\seealso{
\code{\link{read.gulf.spatial}}
}
