% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read}
\alias{read}
\alias{read.gulf.spatial}
\title{Read Spatial Data}
\usage{
read.gulf.spatial(layer, ...)
}
\arguments{
\item{layer}{Character string specifying the layer to be read.}

\item{...}{Other arguments passed onto \code{\link{locate.gulf.spatial}}.}
}
\description{
Read southern Gulf of Saint Lawrence spatial data files.
}
\examples{
read.gulf.spatial(layer = "bounds", survey = "rv")   # Read September survey boundaries.
read.gulf.spatial(layer = "bounds", survey = "sc")   # Read snow crab survey boundaries.
read.gulf.spatial(layer = "stations", survey = "rv") # Read September survey stations.
read.gulf.spatial(layer = "stations", survey = "ns") # Read Northumberland Strait survey stations.
read.gulf.spatial("port")                            # Coordinate port coordinates.

}
\seealso{
\code{\link{locate.gulf.spatial}}
}
