% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scsset.R
\name{longitude.scsset}
\alias{longitude.scsset}
\alias{latitude.scsset}
\title{\strong{\code{scsset}} Class Spatial Functions}
\usage{
\method{longitude}{scsset}(x)

\method{latitude}{scsset}(x)
}
\arguments{
\item{x}{\code{scsset} object.}
}
\description{
Spatial functions for the snow crab set \code{scsset} data class.
}
\section{Methods (by class)}{
\itemize{
\item \code{scsset}: Convert numeric longitude coordinates from \code{scsset} objects to decimal degrees.

\item \code{scsset}: Convert numeric latitude coordinates from \code{scsset} objects to decimal degrees.
}}

\examples{
x <- read.scsset(2020)
longitude(x)

plot(lon(x), lat(x))

}
